requirements.txt
flask==3.0.2
flask-cors==4.0.0
rembg==2.0.64
Pillow==10.2.0
requests==2.31.0 
gunicorn==21.2.0